upload: Clean up error messages from pre-upload scripts
When a pre-upload script prints to stdout/stderr, it would print
immediately after the "Running pre-upload command" line (without a
line break or space). If the command failed, the resulting console
error message was empty (resulting in a blank 'E:' line).

Fix both of these by capturing output from the pre-upload script and
stuff it into the exception. When the exception is caught later, the
stdout/stderr will be printed to the console separately with the 'E:'
prefix.

This has the side-effect of suppressing output when the script is
successful.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>