patchsets: Make virtual diff target with shared history
Previously we just made the first virtual diff target with no parent
resulting in a branch with no shared history with any other
branch. Strictly this shouldn't matter as we only use the
branch for diffs, but we've seen a few cases where pushing this
branch is specifically causing git push to take an incredibly
long time. I don't have a great way to explain this as the virtual
diff target should only be pushing the tree itself, all blobs within
it reuse blobs that were already present in the repo.

Try to create the target with a shared history from an
arbitrary commit to see if this resolves the issues we see.

Topic: vritdiff
Reviewers: brian-k