Only raise in parse_topics when uploading
amend will populate topics so that it can use topic names
to locate commits. However in the case that there is format
or usage error in the commit text, we don't want to raise
at this point since it prevents amend from being used to fix
it in the first place.

Topic: raise
Reviewers: brian-k