Add a simple LRU cache to common git cmds
Simple commands such as ref existence, id lookup,
and distance lookup are pretty common and could be
called with the same args at a higher level such that
its difficult to dedupe between calls. To speed things
up, add a lru_cache decorator so that later calls are
looked up. We don't add this to everything since
more complex functions are unlikely to be called with
the same args multiple times.

These calls are mostly pure functions since revup
generally doesn't change the git database. There are
a few exceptions to that, when it runs "reset" for certain
commands. In these cases, we empty all caches just to be safe.

Topic: cache
Reviewers: brian-k