Catch and warn on github timeouts
Its pretty common for us to hit the 10s github
timeout for very large graphql mutations. In pretty
much all of these cases though, the actual changes
in the request did go through, and a subsequent
revup upload will show everything up to date.

To reflect this and make it nicer for users, catch any
github exception that contains 'timeout' and turn it
into a warning instead. This ensures that we show the topic
list. If the user wishes, they can rerun the command to
verify the changes were actually successful.

Topic: githubtimeout
Reviewers: aaron, brian-k