amend: Add more detail to the reflog action
Previously, amend had a customized reflog message, but
it was static and did not have detail on which change was
amended. This made looking through the reflog pretty confusing
if amend was used many times.

Add additional detail to show whether drop/insert was used and
which commit was targeted with the amend. We don't need to
show the new HEAD because this is already in the reflog.