amend: respect git config commit.cleanup
Git commit has several configuration options
for how to strip comment lines out of the commit
text. We generally want to support these because
they affect whether we throw away prefixed lines
that could be important text to the user.

"strip" is just the default way we've been doing it, so
refactor to also support "scissors" which has a different
comment and removal method. "whitespace" and "verbatim"
are simpler strategies that we can support as well, but get
no additional comment

Fixes: #148