upload: Split populate reviews into 2 phases
This sets the function up to allow arbitrary topic
order in the future.

The first pass handles finding the relative topic, but
does not assume it's already populated. The second pass
needs to assume that the relative is already populated
since it uses it, so it needs to be done in topological
order (which is a followup pr).