amend: Correctly strip comments with no trailing newline
Revup's default COMMIT_EDITMSG does not have a trailing newline; as a
result, the regex used for matching comment lines when cleanup_type is
'strip' will not match the last line in the message and can leave a
trailing comment line. Fix this by making the trailing newline
optional in the regex.

Fixes: 24012d1 ("amend: respect git config commit.cleanup")
Signed-off-by: Brian Kubisiak <brian@kubisiak.com>