upload: Fix review graph comment
If there are two PRs in a relative chain where one PR's URL is a
prefix of another (e.g. when PR 24 is relative to PR 2), the review
graph comment will incorrectly modify the second URL by inserting '**'
at the start and in the middle of the URL. This is due to the naive
string replacement checking for any occurance of each URL.

Fix this by searching instead for the URL followed by the review title
and replacing them both at once instead of individually. This is also
slightly more efficient since there are fewer strings allocated.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>