amend: Allow amending with pure renames
If the only staged changes are pure renames, 'git diff --cached
--quiet' will not report any changes. As a result, attempting to
'revup amend' would skip writing a new tree for the commit and the
rename would still be staged rather than added to the amended commit.

Fix this by turning off rename detection when checking for staged
changes in 'revup amend'. This should be slightly more efficient as
well.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>