upload: Query head oid of a PR via commits
The headRefOid field of a PR actually contains the oid of
whatever the tot of that branch name is, which may not be the
commit that is actually associated with the pr if it has been merged.
To get that correct, we have to query the actual commit list.

This handles a corner case in reordering PRs with the repo's merge
setting as rebase. Somehow this causes one of the reordered PRs to
get merged with 0 commmits, which later causes us to not re-make prs.

With this change, we'll always succeed in remaking prs if they are
erroneously closed or merged on a reorder. Avoiding having them get
closed is a follow-up change.

Topic: query1
Reviewers: aaron, brian-k