upload: Push a dummy commit to work around reordering issues
We've had a long running issue where github may mark prs as merged
or closed when trying to reorder relative series. This is caused
by push + update not being atomic (see comments for more details).

To work around this we'll

- detect it in a dumb way, if 2 or more prs had their base changed
this upload. this will result in some false positives, but the only
cost is an additional ~1s push that people likely won't notice, and
will save us from needing to write more complex logic to detect it
correctly.
- push a dummy commit to every single branch initially that prevents
github from marking any of the prs as merged.
- update as normal
- do an extra push to remove the dummy commit

Topic: reorder4
Reviewers: aaron, brian-k
Fixes: #170